requirements.txt

streamlit
pandas
pytesseract
PyMuPDF
Pillow
